# openapi.yaml
openapi: 3.0.3
info:
  title: brain-server API
  description: >
    HTTP API for the brain-server local agent memory / semantic search server
    (Axum + rusqlite + sqlite-vec + model2vec). This spec is hand-written from
    the source in src/main.rs and src/handlers/*. Endpoints whose contract could
    not be fully verified from source are marked with `x-status: unverified`.
    Every response carries `X-Api-Version: <semver>`; legacy routes carry a
    `Deprecation` header per the deprecation policy in API_CONTRACT.md.
  version: 0.9.8
  x-api-version: "0.9.8"
servers:
  - url: http://127.0.0.1:8765
    description: Default local server (override BIND_HOST/BIND_PORT)
paths:
...
  /v1/embeddings:
    post:
      summary: OpenAI-style embeddings
      operationId: embeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  oneOf:
                    - { type: string }
                    - type: array
                      items: { type: string }
                model: { type: string }
      responses:
        "200":
          description: Embeddings
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, example: list }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        object: { type: string }
                        embedding: { type: array, items: { type: number } }
                        index: { type: integer }
                  model: { type: string }
                  usage:
                    type: object
                    properties:
                      prompt_tokens: { type: integer }
                      total_tokens: { type: integer }
...
  /consolidate/apply:
    post:
      summary: Record operator-chosen evidence links (v0.9.8 Evidence)
      operationId: applyConsolidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyRequest'
      responses:
        "200":
          description: Links recorded (rejected links reported, non-fatal)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplyResponse'
...
  /ingest:
    post:
      summary: Plugin ingest (contract stub)
      operationId: ingest
      x-status: unverified
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Wire contract per API_CONTRACT.md; body not verified in source.
      responses:
        "200":
          description: Unverified response shape
  /memory/{id}:
    delete:
      summary: Forget / delete a memory (contract stub)
      operationId: forget
      x-status: unverified
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Unverified response shape
...
components:
  schemas:
...
    ApplyRequest:
      type: object
      description: v0.9.8 Evidence — operator-chosen typed links to record.
      properties:
        links:
          type: array
          items:
            type: object
            properties:
              from_chunk: { type: integer }
              to_chunk: { type: integer }
              kind: { type: string, enum: [supports, supersedes, contradicts, references, derived_from] }
      required: [links]
    ApplyResponse:
      type: object
      properties:
        recorded: { type: integer }
        rejected: { type: array, items: { type: string } }